+++ src/grokcore/annotation/lazy.py
@@ -99,6 +99,19 @@ class LazyAnnotation(Location):
             return default
         return getattr(storage, key, default)
 
+    @property
+    def _p_changed(self):
+        storage = self.storage
+        if storage is None:
+            return False
+        return storage._p_changed
+
+    @_p_changed.setter
+    def _p_changed(self, value):
+        storage = self.storage
+        if storage is not None:
+            storage._p_changed = value
+
 
 @implementer(IAnnotationFactory)
 class LazyAnnotationFactory(object):